split state code into update and control
